test: consolidate and organize relay_setup and relay_schema tests
- Created tests/test_relay_setup.py by consolidating tests from tests/test_relay.py and tests/test_relay_setup_coverage_fix.py.
- Created tests/test_relay_schema.py for dedicated RELAY_SCHEMA unit tests.
- Moved TestCLIIntegration from tests/test_relay.py to tests/test_cli.py.
- Deleted redundant tests/test_relay.py and tests/test_relay_setup_coverage_fix.py.
- Verified 100% test coverage for src/better_code_review_graph/relay_setup.py.
