+++ tests/test_relay_setup.py
@@ -1,4 +1,4 @@
-"""Tests for relay_schema, relay_setup, and credential_state modules."""
+"""Tests for relay_setup module."""
 
 from __future__ import annotations
 
@@ -15,9 +15,7 @@ from better_code_review_graph.relay_setup import (
     ensure_config,
 )
 
-# Test URL used when exercising the create_session path. Matches the
-# MCP_RELAY_URL env var set by the module-level autouse fixture below
-# (no production default per mode-matrix 2.5).
+# Test URL used when exercising the create_session path.
 TEST_RELAY_URL = "https://relay.example.com"
 
 
@@ -32,46 +30,6 @@ def _default_relay_url_env(monkeypatch):
     monkeypatch.setenv("MCP_RELAY_URL", TEST_RELAY_URL)
 
 
-# ---------------------------------------------------------------------------
-# relay_schema
-# ---------------------------------------------------------------------------
-
-
-class TestRelaySchema:
-    def test_server_name(self):
-        assert RELAY_SCHEMA["server"] == "better-code-review-graph"
-
-    def test_display_name(self):
-        assert RELAY_SCHEMA["displayName"] == "Code Review Graph"
-
-    def test_has_fields(self):
-        fields = RELAY_SCHEMA["fields"]
-        assert len(fields) == 4
-
-    def test_field_keys(self):
-        keys = [f["key"] for f in RELAY_SCHEMA["fields"]]
-        assert keys == [
-            "JINA_AI_API_KEY",
-            "GEMINI_API_KEY",
-            "OPENAI_API_KEY",
-            "COHERE_API_KEY",
-        ]
-
-    def test_all_fields_optional(self):
-        for field in RELAY_SCHEMA["fields"]:
-            assert field["required"] is False
-
-    def test_schema_is_valid_typed_dict(self):
-        assert "server" in RELAY_SCHEMA
-        assert "displayName" in RELAY_SCHEMA
-        assert "fields" in RELAY_SCHEMA
-
-
-# ---------------------------------------------------------------------------
-# relay_setup
-# ---------------------------------------------------------------------------
-
-
 class TestRelaySetupConstants:
     def test_server_name(self):
         assert SERVER_NAME == "better-code-review-graph"
@@ -81,10 +39,6 @@ class TestRelaySetupConstants:
         assert "GEMINI_API_KEY" in CLOUD_KEYS
         assert "OPENAI_API_KEY" in CLOUD_KEYS
         assert "COHERE_API_KEY" in CLOUD_KEYS
-
-    # DEFAULT_RELAY_URL removed per mode-matrix 2.5 (no centralized subdomain
-    # for default-local servers). See test_missing_relay_url_raises below for
-    # the enforcement test.
 
 
 class TestEnsureConfigEnvVar:
@@ -283,11 +237,6 @@ class TestEnsureConfigRelay:
                 assert result is None
 
 
-# ---------------------------------------------------------------------------
-# CLI integration -- tests serve_main calls resolve_credential_state
-# ---------------------------------------------------------------------------
-
-
 class TestApplyConfig:
     def test_apply_config_sets_env(self, monkeypatch):
         for key in CLOUD_KEYS:
@@ -399,32 +348,87 @@ class TestEnsureConfigRelayNotifyFailure:
                 assert result is None
 
 
-class TestCLIIntegration:
-    def test_main_runs_stdio_directly(self):
-        """serve_main(stdio) invokes FastMCP stdio directly (no bridge layer)."""
-        from better_code_review_graph import server as server_module
+# ---------------------------------------------------------------------------
+# Extra coverage from test_relay_setup_coverage_fix.py
+# ---------------------------------------------------------------------------
 
-        with (
-            patch.object(server_module.mcp, "run") as mock_run,
-            patch.dict(os.environ, {"MCP_TRANSPORT": "stdio"}),
+
+@pytest.mark.asyncio
+async def test_read_config_exception_constructor(monkeypatch):
+    """Cover the case where PerPluginStore constructor itself raises."""
+    for key in CLOUD_KEYS:
+        monkeypatch.delenv(key, raising=False)
+
+    # PerPluginStore constructor raises Exception -- should fall through to relay
+    with patch(
+        "better_code_review_graph.relay_setup.PerPluginStore",
+        side_effect=Exception("Disk error"),
+    ):
+        # To avoid going into relay setup, we also mock create_session to fail
+        with patch(
+            "mcp_core.relay.client.create_session",
+            side_effect=Exception("no relay"),
         ):
-            from better_code_review_graph.cli import main
+            result = await ensure_config()
+            assert result is None
 
-            main()
-            mock_run.assert_called_once_with(transport="stdio")
 
-    def test_main_continues_on_relay_error(self):
-        with (
-            patch(
-                "better_code_review_graph.credential_state.resolve_credential_state",
-                side_effect=Exception("relay broken"),
-            ),
-            patch("better_code_review_graph.server.mcp"),
-            patch.dict(os.environ, {"MCP_TRANSPORT": "stdio"}),
+@pytest.mark.asyncio
+async def test_httpx_post_exception_direct(monkeypatch):
+    """Direct port of test_httpx_post_exception from coverage fix file."""
+    for key in CLOUD_KEYS:
+        monkeypatch.delenv(key, raising=False)
+
+    mock_session = MagicMock()
+    mock_session.relay_url = "https://relay.example.com/setup"
+    mock_session.session_id = "test-session"
+
+    mock_client = MagicMock()
+    mock_client.post = AsyncMock(side_effect=Exception("Network error"))
+    mock_client.__aenter__ = AsyncMock(return_value=mock_client)
+    mock_client.__aexit__ = AsyncMock(return_value=False)
+
+    with patch("better_code_review_graph.relay_setup.PerPluginStore") as mock_store_cls:
+        mock_store_cls.return_value.load.return_value = None
+        with patch(
+            "mcp_core.relay.client.create_session",
+            new_callable=AsyncMock,
+            return_value=mock_session,
         ):
-            from better_code_review_graph.cli import main
+            with patch(
+                "mcp_core.relay.client.poll_for_result",
+                new_callable=AsyncMock,
+                return_value={"GEMINI_API_KEY": "new-key"},
+            ):
+                with patch("httpx.AsyncClient", return_value=mock_client):
+                    result = await ensure_config()
+                    assert result is not None
+                    assert result["GEMINI_API_KEY"] == "new-key"
+                    assert os.environ.get("GEMINI_API_KEY") == "new-key"
+    monkeypatch.delenv("GEMINI_API_KEY", raising=False)
 
-            try:
-                main()
-            except Exception:
-                pass
+
+@pytest.mark.asyncio
+async def test_poll_for_result_runtime_error_unexpected_direct(monkeypatch):
+    """Direct port of test_poll_for_result_runtime_error_unexpected."""
+    for key in CLOUD_KEYS:
+        monkeypatch.delenv(key, raising=False)
+
+    with patch("better_code_review_graph.relay_setup.PerPluginStore") as mock_store_cls:
+        mock_store_cls.return_value.load.return_value = None
+        mock_session = MagicMock()
+        mock_session.relay_url = "https://relay.example.com/setup"
+
+        with patch(
+            "mcp_core.relay.client.create_session",
+            new_callable=AsyncMock,
+            return_value=mock_session,
+        ):
+            # RuntimeError with unexpected message -- should return None
+            with patch(
+                "mcp_core.relay.client.poll_for_result",
+                new_callable=AsyncMock,
+                side_effect=RuntimeError("something went wrong"),
+            ):
+                result = await ensure_config()
+                assert result is None
